# Optimization for std.array.Appender