This change brings the performance of appender to the relm of https://github.com/dlang/phobos/pull/5115 in all cases.

DMD for some reason is not able to inline the trusted function, making this almost twice as slow as it would otherwise be.

Benchmark using long ranges of `int`s:

```
# before
3 secs, 91 ms, 452 μs, and 6 hnsecs

# after
1 sec, 907 ms, 67 μs, and 1 hnsec

# after without trusted anonymous function
1 sec, 126 ms, 67 μs, and 4 hnsecs
```