@JackStouffer 2017-02-20
TBQH emplaceRef should be inlined and optimized by DMD to do this already, but this is so commonly used in Phobos that I think it warrants a manual fix