@jmdavis 2017-02-20
I agree with David. This sort of thing gets ugly quickly and reduces maintainability. It may be worth it in some cases, but this is clearly a case of dmd needing to be improved, and anyone who really cares about performance on that level is going to use ldc anyway.

At minimum, if we do merge something like this here, I think that the original code should be left in as a comment with it being made very clear that once dmd is sufficiently improved, the original code will be restored. But even then, I favor not doing this and instead creating a bug report for dmd about this so that the dmd developers have a concrete example of what needs to be improved. And if dmd is improved such that this change wouldn't be necessary, then we'll get far greater improvements across the board.