@JackStouffer 2017-02-20
https://issues.dlang.org/show_bug.cgi?id=17210